+++ Documentation/SBV/Examples/Uninterpreted/AUF.hs
@@ -62,14 +62,3 @@ proveSArray = prove $ do
                 y <- free "y"
                 a :: SArray Word32 Word32 <- newArray_ Nothing
                 return $ thm x y a
-
--- | Prove it using SBV's internal functional arrays.
---
--- >>> proveSFunArray
--- Q.E.D.
-proveSFunArray :: IO ThmResult
-proveSFunArray = prove $ do
-                x <- free "x"
-                y <- free "y"
-                a :: SFunArray Word32 Word32 <- newArray_ Nothing
-                return $ thm x y a