+++ SBVBenchSuite/BenchSuite/Uninterpreted/AUF.hs
@@ -23,13 +23,8 @@ import BenchSuite.Bench.Bench
 benchmarks :: Runner
 benchmarks = rGroup
   [ run "SArray" array `using` runner proveWith
-  , run "SFunArray" funArray `using` runner proveWith
   ]
   where array = do x <- free "x"
                    y <- free "y"
                    a :: SArray Word32 Word32 <- newArray_ Nothing
                    return $ thm x y a
-        funArray = do x <- free "x"
-                      y <- free "y"
-                      a :: SFunArray Word32 Word32 <- newArray_ Nothing
-                      return $ thm x y a